Plugins vim-easytags and tagbar
These give you the power to see what kinds of methods, variables,
functions, and other types of declarations you have in your files. It
relies on a program called ctags being installed, and normally you have
to run ctags to generate the `tags` file for the plugin to work.
Luckily, there is another plugin that makes generating this tags file
easy: vim-easytags. It automatically generates the tags file in the
background whenever you save, so you never have to worry about the tags
file getting out of date.

If you're wondering why `tags` files start showing up in random places,
this is it.