Plugin ctrlp
ctrlp is a magical Vim plugin for finding files quickly. It lets you
"fuzzy find" files, which you'll be familiar with if you use Sublime a
lot. Basically, you can type a subset of the letters in a file's name,
and the plugin will be smart enough to figure it out.