Plugins vim-gitgutter and vim-fugitive
These plugins make working with Git a sinch.

vim-gitgutter shows a +/-/~ next to lines that have been added, removed,
and modified, and also shows a summary of the number of lines affected
in vim-airline's statusbar.

vim-fugitive makes working with Git from within Vim incredibly easy.
Here's a lit of common commands and their analogs to the normal git
commands:

- git add                  --> :Gwrite
- git commit               --> :Gcommit
- git push                 --> :Gpush
- git checkout <file name> --> :Gread
- git blame                --> :Gblame
- ... many more!